chapter_8/requirements.txt
absl_py==1.3.0
ale_py==0.7.5
gym==0.25.2
AutoROM==0.4.2
AutoROM.accept-rom-license==0.4.2
matplotlib==3.7.0
numpy==1.24.3
opencv_python==4.7.0.72
python-snappy==0.6.1
pandas==1.5.3
Pillow==9.5.0
torch==2.0.1
